backend/requirements.txt
affine==2.4.0
amqp==5.1.1
asgiref==3.7.2
aspose-zip==24.9.0
attrs==23.2.0
billiard==4.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dbfread==2.0.7
Django==4.2.5
django-auditlog==2.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
drf_base64==2.0
env==0.1.0
Figures==0.4.4
fonttools==4.43.1
GDAL @ file:///C:/Users/admin/Downloads/GDAL-3.8.4-cp310-cp310-win_amd64.whl#sha256=af9b1fbded4e6e93a0b26e268562514c5be3cd243f957c569ff1d44fba1bcef3
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.5
kombu==5.3.2
lxml==5.1.0
matplotlib==3.8.0
networkx==3.4.2
numpy==1.26.0
packaging==23.2
pandas==2.2.3
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
Pygments==2.16.1
PyJWT==2.8.0
pyogrio==0.10.0
pyparsing==3.1.1
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
rarfile==4.2
rasterio==1.3.10
requests==2.32.3
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.2
vine==5.0.0
wcwidth==0.2.8
wfastcgi==3.0.0
xmljson==0.2.1
